historyService, create contract, inactive contracts
The history Table now only shows the history of a single object, including the username/email of who performed the action. a red error on create contracts now prevents the user from submitting the form, blue warning doesn't. contract_number is now automatically filled in.
If a contract is past its end date, it will be set to inactive